backend/pyproject.toml
[project]
name = "romulus"
version = "0.1.0"
description = "FastAPI application"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
]

[tool.setuptools]
packages = ["game", "game.scene", "game.component", "core"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "py65>=1.1.0",
    "ruff>=0.6.0",
]

[tool.ruff]
# Set the maximum line length
line-length = 120

# Target Python 3.13
target-version = "py313"

# Exclude common directories
extend-exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".pytest_cache",
    "alembic/versions",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults (FastAPI Depends)
    "UP007", # use X | Y for type annotations (not always clearer)
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Use spaces for indentation
indent-style = "space"

# Respect magic trailing commas
skip-magic-trailing-comma = false
